requirements.txt
pandas>=2.0.0
numpy>=1.24.0
great_expectations~=0.18.0
PyYAML == 6.0.1
multiprocess>=0.70.0
colorlog>=6.7.0
tqdm>=4.66.0
pytz>=2023.3
memory-profiler>=0.61.0
orjson>=3.9.0